Refactor the runner a bit
