+++ pkg/skaffold/runner/run.go
@@ -1,0 +1,36 @@
+/*
+Copyright 2019 The Skaffold Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package runner
+
+import (
+	"context"
+	"io"
+
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
+)
+
+// Run builds artifacts, runs tests on built artifacts, and then deploys them.
+func (r *SkaffoldRunner) Run(ctx context.Context, out io.Writer, artifacts []*latest.Artifact) error {
+	if err := r.buildTestDeploy(ctx, out, artifacts); err != nil {
+		return err
+	}
+	if r.runCtx.Opts.Tail {
+		logger := r.newLogger(out, artifacts)
+		return r.TailLogs(ctx, out, logger)
+	}
+	return nil
+}
